Add "sort by source" options to the selection screen
This adds "Source (A-Z)" and "Source (Z-A)" options to the selection screen.

Some normalization is performed on the sources during sorting.
Specifically:
- Diacritics are normalized
- Parentheticals are removed
- Common stop words ("a", "an", "and", "the") are removed
- Any remaining punctuation and whitespace are removed

This should help fix situations such as "Legend of Zelda" vs. "The Legend of Zelda"
and the four different variations on "Pokémon Black and White" used by
the four different characters from that source.