Pot of Greed: Prevent deck overdraft
Intentionally only allows the Random strategy to pick up to 5 cards to trade, instead of up to PoG's max 7.

If all 5 players draw 5 and then choose to trade in 5, they use up 50 cards total. PoG draws two extra for 52, and thus the deck is already empty, so PoG trading in more than the max of 5 would break the game.

This is exceedingly unlikely, but the possibility has been patched out anyway.