Pot of Greed: improve AI
Adds a bit more awareness of the extra possibilities for hands.

For example, previously upon receiving a five-card flush, the AI would not trade, as there's nothing to do to improve a five-card flush... *if* you only have five cards.

Now the AI will trade in the other two cards in hopes of upgrading to a six-card or even seven-card flush.

This makes Pot of Greed an even more difficult opponent, but it also increases the probability of it getting one of the wacky made-up hands, thus enhancing the joke.