Pot of Greed: fix dialogue that extends past speech bubble
This is a pretty glaring flaw, and yet no one has reported it in two years, implying that no one has seen this line.

And it's the single line for defeating him, which in turn implies that not a single person has ever defeated Pot of Greed!

(Obviously it's just that people tend not to report even the most egregious errors. But I want to believe.)