Add sprite clipping properties for epilogues
This adds 5 animateable properties that allow sprites to be partially clipped:
- `clipleft`
- `clipright`
- `clipbottom`
- `cliptop`
- `clipradius`

Together, these properties define a (potentially rounded) rectangular
region that sets which part of a sprite should be shown: areas
inside the rectangle are displayed, while areas outside are hidden.
This applies to both the sprite itself, as well as any child sprites.

To be specific: `clipleft`, `clipright`, `cliptop`, and `clipbottom`
define an inset rectangle whose sides are offset from the edges of the
sprite by a given number of pixels. The `clipradius` property can be
used to round the corners of the rectangle using a radius given in pixels.

All five of these properties can be animated using keyframes.

One example use-case for this feature is to create "side panels"
containing animated content that must be constrained to fit within the
bounds of the panel. Clipping regions make this easy: one can simply
parent all of the sprites within the "side panel" to the sprite for
the panel's frame and set a clip region for it; any sprites (or parts thereof)
that go outside of the panel will then be automatically hidden.