Panty: Saved in v6.8.2.6
Testing an experimental version can lead to unexpected consequences. I have attempt to untangle what was new and what was lost on saving, and the only thing I know for sure that was messed up was the ID numbers, which are still assigned to the lines but the new ones were not recorded in editor.xml. Bug report submitted.