Pose Matrix: remove part-removing template logic
Disables logic within the Pose Matrix templating code that removes
placeable items present in the pose code in certain cases, which should
make the templating logic more predictable and easier to use in the long run.

This is technically a breaking change, but I'm not sure if any existing
pose matrices depend on this behavior.