Feature characters on their birthdays
Birthdays are specified in meta.xml like `<birthday month="6" day="10"/>`.

A character whose birthday it is (based on the local month and day when the game is loaded) will have top priority to be featured as a preselect (because it's only one day and otherwise characters whose birthdays occur during an event will be at a disadvantage) with a balloon above their head.

They will not be sorted before event characters on the individual select screen, nor given their own partition, but be sorted before the rest of the ordinary characters.