Major change to make ~clothing.reveals~ work, basically.
We now don't remove stripped articles from the wardrobe, only mark them as removed. This makes it possible to go back and see the `Player`'s clothing state before the last strip.

We probably don't need `Player.prototype.isCovered()`, `getRevealedPosition()`, or even `getClothingTrigger()` to take the removed article as an argument, since we always set the `Player`'s `removedClothing` to it.