Handle structured credits
Expects one or more `<artist>` and `<writer>` elements, as well as `<othercred>` (but that isn't used yet), with `comment` and `minor` attributes.

If there's only one element of a certain type and it doesn't have any attributes, it's treated as legacy credits and parsed. It should work in almost all cases.