Nagisa: Birthday update!
Replies to some Hand lines from Jura

Also, in Nagisa's conversations where Nagisa must strip, the natural priority for lines to characters on the left of the screen is 151, whereas to the right of the screen it's 150 (due to no marker check). This could result in her favoring one character when she herself strips, even if she had conversations with characters on both sides. For most of these lines, I've now set the manual priority of these lines to 150 so that she'll be able to pick conversation partners without a bias to one side. Once a conversation is underway, the unchanged natural priority of the subsequent lines will keep it going.

Also, happy Christmas Eve, Nagisa!