CI/CD: transpile JS with Babel during deployment
This adds a Babel transpilation step to the deployment process,
which will automatically rewrite core game Javascript to only
use widely-supported syntax features. This only affects the online
version; offline users are expected to use modern browsers that
support the syntax features we use as-is.

The transpilation target I've set here applies to all browsers
with more than 0.1% usage, which should in total cover
about 94.2% of users worldwide (see https://browsersl.ist/#q=%3E+0.1%25).

This transpilation step only modifies syntax, and does not include
polyfills needed for full backwards-compatibility.

This will require Node.js to be installed on the production server
as well as the infrastructure server (for testing deployments).