Add core-js for polyfills
This adds a static copy of core-js v3.35.0 to provide polyfills for
the entire JS standard library, complementing the JS transpilation step
and (hopefully) making our JS code fully backwards compatible with pretty
much every browser in use today.

For right now, this just pulls in the entirety of CoreJS as a bundle;
Babel could theoretically emit specific imports for individual features
that need to be polyfilled, but getting our code to work with bundlers
like Browserify is hard because our code doesn't use modules.